Protein ontogenies (#827)
* - Replace columns "Protein" and "Ontogeny" by the column "Protein Ontogenies" in the files "Individuals" and "Populations"
- Mention in "NEWS"

* - Update vignette

* rotein ontogenies for individuals and populations are now defined in one column
`Protein Ontogenies` in the sheets `IndividualBiometrics` (for individuals) or `Demographics`
(for populations). The columns `Protein` and `Ontogeny` are no longer supported. The new column
`Protein Ontogenies` is a comma-separated list of protein names and ontogeny names pairs. For example:
`CYP3A4:CYP3A4,CYP2D6:CYP2C8` will create a CYP3A4 ontogeny for the protein CYP3A4 and a CYP2D6 ontogeny for the protein CYP2C8.

* Styler

* - Fix iteration
- Update test excel

* - Remove base packages from IMPORTS
- Remove ospsuite.utils from IMPORTS, as it is also stated in DEPENDS
- Added a test

* - Added testing for correct ontogenies